Name a start pose with the rig it belongs to
`main_cfg` bound the Franka nominal joints while binding no arm, so the
documented `--robot_arm=@...kinova` override kept them: Kinova takes seven
joints too, so the right stick drove it to a Franka pose with nothing raised.
An arm named without its joints is now refused up front.

`real` checks at startup that what an episode readies is what the rig has, so
overriding its embodiment no longer waits until the first press to fail.

`IdleSession` joins the other shared test doubles rather than standing twice.
