+++ positronic/tests/testing_coutils.py
@@ -114,3 +114,24 @@ def run_scripted_agent(
     driver = ManualDriver(script=script)
     scheduler = world.start([agent, driver])
     drive_scheduler(scheduler, steps=steps)
+
+
+class IdleSession:
+    """A policy session that records what it is shown and places nothing.
+
+    Its policy owns an ``observations`` list, which is where each one lands.
+    """
+
+    def __init__(self, policy):
+        self._policy = policy
+
+    def __call__(self, obs):
+        self._policy.observations.append(obs)
+        return []
+
+    @property
+    def meta(self):
+        return {}
+
+    def close(self):
+        pass
